[server] Moved QgsBufferServerRequest into its own header and cpp file
